1. Modified compiler to have generated setXXX methods return this so that data structure construction calls can be chained in a builder like pattern.
2. Made ByteBufferInputStream derive from FSInputStream so that it
becomes seekable (inherently supported by ByteBuffer anyhow).

3. Made MMapFileInputStream derive from FSInputStream and made
MMapFile.newInputStream return an FSDataInputStream, so that the
mmap'd file stream can be interchanged with a HDFS stream in
certain parts of the codebase.

4. Reverted some changes in the RiceCoder for now, and also
added a constructor for RiceCodeReader that takes an
FSDataInputStream, thus enabling a MMapFileStream to be used
to initialize a Reader object. This eliminated a buffer copy
from an FSDataInputStream to ByteBuffer (the other type allowed
in the Reader's constructor).

5. Added a hasMoreData method to TFileReader to enable an
after 'next' check to see if EOF condition has been hit. This
is possible because the last call to TFile.Scanner's next sets
it in an EOF state which can be checked via its atEnd method.